Check the Accept header and return 406 if we cannot abide.
