+++ imgfac/rest/RESTtools.py
@@ -15,6 +15,7 @@
 
 import logging
 from imgfac.rest.bottle import *
+from imgfac.ApplicationConfiguration import ApplicationConfiguration
 
 log = logging.getLogger(__name__)
 
@@ -39,3 +40,25 @@ def form_data_for_content_type(content_type):
     except Exception as e:
         log.exception(e)
         raise HTTPResponse(status=500, output=e)
+
+def log_request(f):
+    def decorated_function(*args, **kwargs):
+        if(ApplicationConfiguration().configuration['debug']):
+            log.debug('Handling %s HTTP %s REQUEST for resource at %s: %s' % (request.headers.get('Content-Type'),
+                                                                              request.method,
+                                                                              request.path,
+                                                                              request.body.read()))
+        return f(*args, **kwargs)
+    decorated_function.__name__ = f.__name__
+    return decorated_function
+
+def check_accept_header(f):
+    def decorated_function(*args, **kwargs):
+        accept_header = request.get_header('Accept', None)
+        if(accept_header and ('application/json' not in accept_header)):
+            log.debug('Returning HTTP 406, unsupported response type: %s' % accept_header)
+            raise HTTPResponse(status=406, output='Responses in %s are currently unsupported. Please try application/json or remove the Accept header from the request.' % accept_header)
+        else:
+            return f(*args, **kwargs)
+    decorated_function.__name__ = f.__name__
+    return decorated_function
